backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
graphene-django==3.2.3
django-filter==25.2
django-cors-headers==4.9.0
python-dotenv==1.2.1
requests==2.32.5
PyMySQL==1.1.2

# Prod / déploiement (Render)
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
psycopg[binary]==3.2.10


